# docs/03-api/OPENAPI.yaml
openapi: 3.0.3
info:
  title: API Automation Testing Platform
  version: 0.1.0
  description: API 自动化测试平台 MVP 接口契约草案
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    ProjectCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
    EnvironmentCreate:
      type: object
      required: [name, base_url]
      properties:
        name:
          type: string
        base_url:
          type: string
        headers:
          type: object
          additionalProperties: true
        variables:
          type: object
          additionalProperties: true
        is_default:
          type: boolean
    CollectionCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
    TestCaseCreate:
      type: object
      required: [name, method, path]
      properties:
        name:
          type: string
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        path:
          type: string
        headers:
          type: object
          additionalProperties: true
        query_params:
          type: object
          additionalProperties: true
        body_type:
          type: string
          enum: [none, json, form, raw]
        body:
          nullable: true
        assertions:
          type: array
          items:
            type: object
            additionalProperties: true
        timeout_seconds:
          type: integer
          default: 30
    RunCreate:
      type: object
      required: [environment_id, scope]
      properties:
        name:
          type: string
        environment_id:
          type: string
          format: uuid
        scope:
          type: string
          enum: [case, collection, project]
        collection_id:
          type: string
          format: uuid
          nullable: true
        case_id:
          type: string
          format: uuid
          nullable: true
    # ---- F012 + F013 ---------------------------------------------------------
    OpenApiImportDocument:
      type: object
      description: F013 单个 OpenAPI 文档引用（request body of documents[] element）
      properties:
        source_url:
          type: string
          format: uri
          description: OpenAPI 文档 URL（与 source_content 二选一）
        source_content:
          type: object
          description: OpenAPI 文档原文（与 source_url 二选一）
          additionalProperties: true
        tags:
          type: array
          items: { type: string }
          description: 可选，按 tag 过滤 operation
        name_prefix:
          type: string
          maxLength: 80
    OpenApiImportRequest:
      type: object
      description: F012/F013 OpenAPI 导入请求体。F013 批量通过 documents[] 字段启用；与单文档字段互斥。
      properties:
        source_url:
          type: string
          format: uri
        source_content:
          type: object
          additionalProperties: true
        tags:
          type: array
          items: { type: string }
        on_conflict:
          type: string
          enum: [skip, overwrite]
          default: skip
        dry_run:
          type: boolean
          default: true
        name_prefix:
          type: string
          maxLength: 80
        documents:
          type: array
          description: F013 批量：1~5 个 OpenAPI 文档引用（上限配置项 OPENAPI_BATCH_MAX_DOCS）
          items:
            $ref: '#/components/schemas/OpenApiImportDocument'
    BatchImportPreviewResponse:
      type: object
      description: F013 批量导入预览（dry_run=true）
      properties:
        suite_id: { type: string, format: uuid }
        suite_name: { type: string }
        total_documents: { type: integer }
        total_operations: { type: integer }
        documents:
          type: array
          items:
            type: object
            properties:
              doc_index: { type: integer }
              source: { type: string }
              spec_version: { type: string }
              base_path: { type: string }
              total: { type: integer }
              new_count: { type: integer }
              existing_count: { type: integer }
              skipped_count: { type: integer }
              operations:
                type: array
                items: { type: object, additionalProperties: true }
              errors:
                type: array
                items: { type: string }
        errors:
          type: array
          items: { type: string }
    BatchImportResponse:
      type: object
      description: F013 批量导入提交结果（dry_run=false&preview_id=...&batch=true）
      properties:
        total_documents: { type: integer }
        total_attempted: { type: integer }
        total_succeeded: { type: integer }
        documents:
          type: array
          items:
            type: object
            properties:
              doc_index: { type: integer }
              source: { type: string }
              created: { type: array, items: { type: string } }
              skipped: { type: array, items: { type: string } }
              overwritten: { type: array, items: { type: string } }
              errors: { type: array, items: { type: string } }
        errors:
          type: array
          items: { type: string }
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: 刷新 Token
      responses:
        '200':
          description: 刷新成功
  /auth/logout:
    post:
      tags: [Auth]
      summary: 退出登录
      responses:
        '200':
          description: 退出成功
  /auth/me:
    get:
      tags: [Auth]
      summary: 当前用户信息
      responses:
        '200':
          description: 成功
  /users:
    get:
      tags: [User]
      summary: 用户列表
      responses:
        '200':
          description: 成功
  /users/{user_id}:
    get:
      tags: [User]
      summary: 用户详情
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
    put:
      tags: [User]
      summary: 更新用户
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
    delete:
      tags: [User]
      summary: 禁用用户
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
  /projects:
    post:
      tags: [Project]
      summary: 创建项目
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: 创建成功
    get:
      tags: [Project]
      summary: 项目列表
      responses:
        '200':
          description: 成功
  /projects/{project_id}:
    get:
      tags: [Project]
      summary: 项目详情
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
    put:
      tags: [Project]
      summary: 更新项目
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
    delete:
      tags: [Project]
      summary: 删除项目
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 删除成功
  /projects/{project_id}/environments:
    post:
      tags: [Environment]
      summary: 创建环境
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentCreate'
      responses:
        '201':
          description: 创建成功
    get:
      tags: [Environment]
      summary: 环境列表
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
  /projects/{project_id}/collections:
    post:
      tags: [Collection]
      summary: 创建集合
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreate'
      responses:
        '201':
          description: 创建成功
    get:
      tags: [Collection]
      summary: 集合列表
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
  /collections/{collection_id}/cases:
    post:
      tags: [TestCase]
      summary: 创建用例
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCaseCreate'
      responses:
        '201':
          description: 创建成功
    get:
      tags: [TestCase]
      summary: 用例列表
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
  /cases/{case_id}/run:
    post:
      tags: [Run]
      summary: 执行单用例
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: 执行完成
  /projects/{project_id}/runs:
    post:
      tags: [Run]
      summary: 创建执行批次（F014：可传 ?concurrency= 覆盖默认并发度）
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: concurrency
          in: query
          required: false
          description: |
            F014：单 Run 内同时在飞的最大 case 数（1 ≤ N ≤ 64）。
            不传则使用服务器默认 settings.TEST_RUN_MAX_CONCURRENCY（默认 4）。
            传 1 恢复串行执行。越界返回 422。
          schema:
            type: integer
            minimum: 1
            maximum: 64
            default: null
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreate'
      responses:
        '201':
          description: 创建成功
        '422':
          description: concurrency 越界 / scope 为空 / 请求体验证失败
    get:
      tags: [Run]
      summary: 执行历史
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
  /runs/{run_id}:
    get:
      tags: [Run]
      summary: 执行详情
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
  /runs/{run_id}/results:
    get:
      tags: [Run]
      summary: 执行结果列表
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
  /results/{result_id}:
    get:
      tags: [Run]
      summary: 单条结果详情
      parameters:
        - name: result_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
  # ---- F012 + F013 -----------------------------------------------------------
  /projects/{project_id}/suites/{suite_id}/import/openapi:
    post:
      tags: [OpenAPI]
      summary: F012 单文档 / F013 多文档批量 OpenAPI 导入（?batch=true 启用批量）
      parameters:
        - name: project_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: suite_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: batch
          in: query
          required: false
          description: F013 批量开关；true=启用 documents[]，false=走 F012 单文档语义
          schema: { type: boolean, default: false }
        - name: dry_run
          in: query
          required: false
          schema: { type: boolean, default: true }
        - name: preview_id
          in: query
          required: false
          description: dry_run=false 时需提供首次预览返回的 preview_id
          schema: { type: string }
        - name: on_conflict
          in: query
          required: false
          schema: { type: string, enum: [skip, overwrite], default: skip }
        - name: name_prefix
          in: query
          required: false
          schema: { type: string, maxLength: 80 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenApiImportRequest'
      responses:
        '200':
          description: 成功（F012 单文档：ImportPreviewResponse / ImportResponse；F013 批量：BatchImportPreviewResponse / BatchImportResponse）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 导入过程错误（含 OPENAPI_PARSE_ERROR / OPENAPI_FETCH_ERROR / OPENAPI_IMPORT_CONFLICT / OPENAPI_BATCH_LIMIT_EXCEEDED）
        '401': { description: 未认证 }
        '403': { description: 非项目 owner / admin }
        '404': { description: 项目或套件不存在 }
        '422':
          description: 参数校验失败（含 documents[] 越界、URL scheme 非法、单文档字段与 documents[] 同传等）